Add files via upload
Getting some errors with self as a parameter. We've been narrowing it down, but it thinks we are giving an extra argument. Ex: 
"error, 2 arguments requested (3 given)"
This is in Docent. In terminal of US it works to declare as sensor = US(20, 21)
The init command didn't seem to work either